+++ cmd/in.go
@@ -47,7 +47,7 @@ var inCmd = &cobra.Command{
 		var err error
 
 		tei := dto.TimeEntryImpl{
-			WorkspaceID:  viper.GetString("workspace"),
+			WorkspaceID:  viper.GetString(WORKSPACE),
 			TagIDs:       tags,
 			TimeInterval: dto.TimeInterval{},
 			Billable:     !notBillable,
@@ -78,7 +78,7 @@ var inCmd = &cobra.Command{
 
 		format, _ := cmd.Flags().GetString("format")
 		asJSON, _ := cmd.Flags().GetBool("json")
-		return newEntry(c, tei, viper.GetBool("interactive"), viper.GetBool("allow-project-name"), true, format, asJSON)
+		return newEntry(c, tei, viper.GetBool(INTERACTIVE), viper.GetBool(ALLOW_PROJECT_NAME), true, format, asJSON)
 	}),
 }